подправил структуру каталогов. чут не угрохал проект)).  сохроняю на случай отката
