+++ CLIENT_FOLDER/dfsg34334f.das.txt
@@ -1,0 +1,3 @@
+esfj;sdjaf;jds;fjsd;lkfkldf
+sdafdshf;ksdjf
+dsfdsljf;dsf
