+++ pom.xml
@@ -9,10 +9,10 @@
     <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
     <modules>
-        <module>clientPart</module>
+
         <module>serverPart</module>
-        <module>srareClass</module>
-        <module>ObjectCreatorClassLib1</module>
+
+        <module>cleintPart</module>
 
     </modules>
